+++ src/events/misc.h
@@ -8,6 +8,7 @@ void	event_serverFeatures(struct irc_message_compo *);
 void	event_channelCreatedWhen(struct irc_message_compo *);
 void	event_channelModeIs(struct irc_message_compo *);
 void	event_channel_forward(struct irc_message_compo *);
+void	event_invalidUsername(struct irc_message_compo *);
 void	event_local_and_global_users(struct irc_message_compo *);
 void	event_nicknameInUse(struct irc_message_compo *);
 void	event_userModeIs(struct irc_message_compo *);